+++ components/themes/ui/card/card-container/CardContainer.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from "vitest";
 import CardContainer from "./CardContainer.vue";
-import { render } from "~/tests/testingLibraryVue";
+import { render } from "~~/tests/testing-library-vue";
 
 describe("CardContainer", () => {
   test("mounts and is still the same", () => {
