+++ pages/settings/templates.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts"></script>
+
+<template>
+  <div>
+    <h1>Settings 360 templates</h1>
+  </div>
+</template>
