+++ plugins/vue-draggable.ts
@@ -1,0 +1,5 @@
+import { VueDraggableNext } from "vue-draggable-next";
+
+export default defineNuxtPlugin((nuxtApp) => {
+  nuxtApp.vueApp.component("draggable", VueDraggableNext);
+});
